pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "transferx"
version = "0.1.0"
description = "TransferX MVP"
requires-python = ">=3.11"
dependencies = [
  "Django~=5.2",
  "psycopg[binary]>=3.1",
  "django-htmx>=1.18",
  "httpx>=0.27",
  "django-ratelimit>=4.1",
  "gunicorn>=22.0",
  "pytest>=8",
  "pytest-django>=4.8",
  "ruff>=0.5",
  "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.format]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
pythonpath = ["src"]
testpaths = ["tests"]
